Update actual value text. Attack still not working.

+++ Assets/Scripts/Mapa/GlobalVariables.cs
@@ -41,15 +41,15 @@ public class GlobalVariables : MonoBehaviour
     public double defense = 3.0;
     public double lucky = 1.0;
 
-    public int life = 100;
-    public int mana = 150;
-    public int stamina = 100;
+    public double life = 100.0;
+    public double mana = 150.0;
+    public double stamina = 100.0;
     
-    public int simpleAttack = 1;
-    public int poisonAttack = 0;
-    public int explosaoArcana = 0;
-    public int lifeOrDeath = 0;
-    public int invocacaoProfana = 0;
+    public double simpleAttack = 1;
+    public double poisonAttack = 0;
+    public double explosaoArcana = 0;
+    public double lifeOrDeath = 0;
+    public double invocacaoProfana = 0;
     
 
     [Header("BackGround Musics")]
